refactor: make error page as reusable component
