+++ client/src/App.tsx
@@ -1,11 +1,11 @@
 
-import HomePage from "./pages/HomePage"
-import NotFoundPage from "./pages/NotFoundPage"
-import { Route, Routes } from "react-router"
 import {
   QueryClient,
   QueryClientProvider
-} from '@tanstack/react-query'
+} from '@tanstack/react-query';
+import { Route, Routes } from "react-router";
+import ErrorPage from "./pages/ErrorPage";
+import HomePage from "./pages/HomePage";
 
 const queryClient = new QueryClient();
 
@@ -15,7 +15,7 @@ function App() {
       <QueryClientProvider client={queryClient}>
         <Routes>
           <Route index element={<HomePage />} />
-          <Route path="*" element={<NotFoundPage />} />
+          <Route path="*" element={<ErrorPage title={"Something's missing."} errorCode={404} />} />
         </Routes>
       </QueryClientProvider>
     </>
