+++ client/src/pages/ErrorPage.tsx
@@ -1,12 +1,17 @@
 import { NavLink } from "react-router"
 
-function NotFoundPage() {
+interface ErrorPageProps {
+    title: string,
+    errorCode: number
+}
+
+function ErrorPage({ title, errorCode }: ErrorPageProps) {
     return (
         <section className="bg-white">
             <div className="py-8 px-4 mx-auto max-w-screen-xl lg:py-16 lg:px-6">
                 <div className="mx-auto max-w-screen-sm text-center">
-                    <h1 className="mb-4 text-7xl tracking-tight font-extrabold lg:text-9xl text-primary-600 dark:text-primary-500">404</h1>
-                    <p className="mb-4 text-3xl tracking-tight font-bold text-gray-900 md:text-4xl dark:text-white">Something's missing.</p>
+                    <h1 className="mb-4 text-7xl tracking-tight font-extrabold lg:text-9xl text-primary-600 dark:text-primary-500">{errorCode}</h1>
+                    <p className="mb-4 text-3xl tracking-tight font-bold text-gray-900 md:text-4xl dark:text-white">{title}</p>
                     <p className="mb-4 text-lg font-light text-gray-500 dark:text-gray-400">Sorry, we can't find that page. You'll find lots to explore on the <NavLink className="text-blue-500 hover:text-blue-800" to={"/"}>home page.</NavLink> </p>
                 </div>
             </div>
@@ -14,4 +19,4 @@ function NotFoundPage() {
     )
 }
 
-export default NotFoundPage
+export default ErrorPage;
